Parsing Code for LIDAR and Odometry Messages
The incomming messages were interpreted as strings.

parse_LIDAR_msg:
Isolates a message from a message string. It isolates the first string that occurs between two "---START---" signals. Afterwards it checks if the "___END___" signal is included. If not the message is not a whole one and it returns false, otherwise true.

isolate_LIDAR_ranges:
This function is for isolating the LIDAR ranges data, which is an array of 360 range values. This values are stored in a double array. The function returns true, if the casting was a success, otherwise false.